Fix issues in catboost/cuda that VS2015 complaints about
MLTOOLS-1680

ref:cee4c9c620d9079148bc0988a20571145a524211

+++ catboost/cuda/methods/leaves_estimation/descent_helpers.h
@@ -95,7 +95,7 @@ namespace NCatboostCuda {
             SolveLinearSystemCholesky(&sigma,
                                       &solution);
 
-            for (uint i = 0; i < solution.size(); ++i) {
+            for (size_t i = 0; i < solution.size(); ++i) {
                 MoveDirection[i] = (float)solution[i];
             }
         }
